refactoring usage of BLUIMA_RESOURCE_DIR
`TestWithBluimaResource` cannot be in a test scope of any module since it is shared among several modules. Hence a new module for test was created and added with test scope in the modules that need this feature in their tests.
