+++ modules/bluima_opennlp/pom.xml
@@ -29,5 +29,12 @@
       <artifactId>jules-utilities</artifactId>
       <version>2.1.1</version>
     </dependency>
+    
+    <dependency>
+      <artifactId>bluima_test_helper</artifactId>
+      <groupId>${project.groupId}</groupId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
